add copleted class and refactoring
