+++ src/components/TodoInfo/TodoInfo.tsx
@@ -1,15 +1,18 @@
-import { Todo } from '../../types';
+import classNames from 'classnames';
+import { TodoWithUser } from '../../types';
+import './TodoInfo.scss';
 import { UserInfo } from '../UserInfo';
 
 type Props = {
-  todo: Todo;
+  todo: TodoWithUser;
 };
 
 export const TodoInfo: React.FC<Props> = ({ todo }) => {
   return (
     <article
       data-id={todo.id}
-      className="TodoInfo TodoInfo--completed"
+      className={classNames('TodoInfo',
+        { 'TodoInfo--completed': todo.completed })}
     >
       <h2 className="TodoInfo__title">
         {todo.title}
