+++ src/components/TodoList/TodoList.tsx
@@ -1,8 +1,8 @@
-import { Todo } from '../../types';
+import { TodoWithUser } from '../../types';
 import { TodoInfo } from '../TodoInfo';
 
 type Props = {
-  todos: Todo[];
+  todos: TodoWithUser[];
 };
 
 export const TodoList: React.FC<Props> = ({ todos }) => {
