+++ src/components/UserInfo/UserInfo.css
@@ -1,0 +1,3 @@
+.UserInfo {
+  font-style: italic;
+}
