+++ src/types/Todo.ts
@@ -1,0 +1,16 @@
+import { User } from './User';
+
+export interface TodoWithoutUser {
+  id: number;
+  title: string;
+  completed: boolean;
+  userId: number;
+}
+
+export interface TodoWithUser {
+  id: number;
+  title: string;
+  completed: boolean;
+  userId: number;
+  user: User | null;
+}
